# lib/api-zod/src/generated/types/attacksResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dir3 / درع — AI Security Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Attack } from "./attack";

export interface AttacksResponse {
  total: number;
  attacks: Attack[];
}