# lib/api-zod/src/generated/types/playgroundRunStarted.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dir3 / درع — AI Security Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaygroundRunStarted {
  run_id: string;
  share_id: string;
  total_attacks: number;
  status: string;
}