# lib/api-zod/src/generated/types/scannerRunRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dir3 / درع — AI Security Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ScannerRunRequestInputType } from "./scannerRunRequestInputType";

export interface ScannerRunRequest {
  input_type?: ScannerRunRequestInputType;
  system_prompt: string;
  frameworks?: string[];
  domain?: string;
}