# lib/api-zod/src/generated/types/shieldGenerateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dir3 / درع — AI Security Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ShieldGenerateResponse {
  shield_id: string;
  input_prompt: string;
  output_prompt: string;
  techniques_applied: string[];
  categories_now_resisted: string[];
  estimated_score_improvement?: number;
  diff_summary_en?: string;
  diff_summary_ar?: string;
}